autotest-results: move helper function and join on the correct table names
